# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Media API",
        "description": "Welcome to the documentation for our Social Media API built using Node.js and Express.js! This API provides a set of endpoints to power your social networking application, allowing users to create profiles, post content, interact with others, and more.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "jwt"
            }
        },
        "schemas": {
...
            "userProfile": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "coverImage": {
                                "type": "object",
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "localPath": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    }
                                }
                            },
                            "firstName": {
                                "type": "string"
                            },
                            "lastName": {
                                "type": "string"
                            },
                            "dob": {
                                "type": "string"
                            },
                            "location": {
                                "type": "string"
                            },
                            "bio": {
                                "type": "string"
                            },
                            "phoneNumber": {
                                "type": "number"
                            },
                            "owner": {
                                "type": "string"
                            },
                            "account": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "responsePost": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "url": {
                                    "type": "string"
                                },
                                "localPath": {
                                    "type": "string"
                                },
                                "_id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "coverImage": {
                                "type": "object",
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "localPath": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    }
                                }
                            },
                            "firstName": {
                                "type": "string"
                            },
                            "lastName": {
                                "type": "string"
                            },
                            "dob": {
                                "type": "string"
                            },
                            "location": {
                                "type": "string"
                            },
                            "bio": {
                                "type": "string"
                            },
                            "phoneNumber": {
                                "type": "number"
                            },
                            "owner": {
                                "type": "string"
                            },
                            "account": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "_id": {
                                    "type": "string"
                                },
                                "text": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "likes": {
                        "type": "number"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Users Profile"
                ],
                "summary": "Get my profile",
                "description": "The API endpoint allows a logged-in user to fetch their social media profile. \n \n  When the user is authenticated and authorized, accessing this endpoint will return the user's social media profile information, such as their username, bio, avatar, cover image, and other relevant details.",
                "responses": {
                    "200": {
                        "description": "Get my profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/userProfile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/profile/": {
            "patch": {
                "tags": [
                    "Users Profile"
                ],
                "summary": "Create or Update user profile",
                "description": "The API endpoint allows users to create or update their social media profiles. \n\n By accessing this endpoint, users can make changes to their profile information such as name, bio, contact details, dob, basic details or any other relevant fields. \n \n NOTE: To update cover image we have separate endpoint. \n\n Update cover image: Apps > Social Media > Update cover image",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "bio": {
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "type": "number"
                                    },
                                    "owner": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated user Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/userProfile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/posts/": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a post",
                "description": "The API allows a logged-in user to create a social media post by providing the necessary details in the request body. \n\n The user can include the content of the post, relevant tags, and attach images (upto 6) as part of the request. \n\n NOTE: As this api requires formData to be used, to send multiple tags client must send them with the respective index.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "array",
                                        "maxItems": 6,
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "tags[0]": {
                                        "type": "string"
                                    },
                                    "tags[1]": {
                                        "type": "string"
                                    },
                                    "tags[2]": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/responsePost"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all users posts",
                "description": "The API endpoint allows users to fetch all posts posted by themselves within the social media application. \n\n By accessing this endpoint, users can retrieve a collection of posts containing relevant information.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "description": "Page number for pagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all users posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "page": {
                                                    "type": "string"
                                                },
                                                "totalPosts": {
                                                    "type": "number"
                                                },
                                                "posts": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/responsePost"
                                                    }
                                                }
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    }
}